Move patch to new eslint package folder
